Changed variables and added components page
